# metric: remove mutex around windowed histogram update