@dhartunian 2025-02-05
> This looks overall pretty good - is there a way to test this behavior at all?

Does this existing test satisfy the requirement? https://github.com/cockroachdb/cockroach/blob/master/pkg/util/metric/metric_test.go#L433-L463